refactor: use microservice architecture
