+++ src/components/Trade/TradeInputV2.tsx
@@ -5,8 +5,8 @@ import { KnownChainIds } from '@shapeshiftoss/types'
 import { useController, useFormContext, useWatch } from 'react-hook-form'
 import { useHistory } from 'react-router'
 import { SlideTransition } from 'components/SlideTransition'
-import { useSwapperService } from 'components/Trade/hooks/useSwapper/useSwapperService'
 import { getSendMaxAmount } from 'components/Trade/hooks/useSwapper/utils'
+import { useSwapperService } from 'components/Trade/services/useSwapperService'
 import { bnOrZero } from 'lib/bignumber/bignumber'
 import { selectFeeAssetById } from 'state/slices/assetsSlice/selectors'
 import { selectPortfolioCryptoBalanceByFilter } from 'state/slices/portfolioSlice/selectors'
