+++ src/components/Trade/services/useTradeQuoteService.ts
@@ -1,0 +1,127 @@
+import { skipToken } from '@reduxjs/toolkit/query'
+import { fromAssetId } from '@shapeshiftoss/caip'
+import { UtxoBaseAdapter } from '@shapeshiftoss/chain-adapters'
+import { GetTradeQuoteInput, UtxoSupportedChainIds } from '@shapeshiftoss/swapper'
+import { KnownChainIds } from '@shapeshiftoss/types'
+import { useEffect, useState } from 'react'
+import { useFormContext, useWatch } from 'react-hook-form'
+import { useSelector } from 'react-redux'
+import {
+  getFirstReceiveAddress,
+  getUtxoParams,
+  isSupportedNoneUtxoSwappingChain,
+  isSupportedUtxoSwappingChain,
+  TradeQuoteInputCommonArgs,
+} from 'components/Trade/hooks/useSwapper/utils'
+import { TradeState } from 'components/Trade/types'
+import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingleton'
+import { useWallet } from 'hooks/useWallet/useWallet'
+import { useGetTradeQuoteQuery } from 'state/apis/swapper/swapperApi'
+import { selectAccountSpecifiers } from 'state/slices/accountSpecifiersSlice/selectors'
+import { selectFiatToUsdRate } from 'state/slices/marketDataSlice/selectors'
+import { useAppSelector } from 'state/store'
+
+export const useTradeAmountService = () => {
+  // Form hooks
+  const { control, setValue } = useFormContext<TradeState<KnownChainIds>>()
+  const sellTradeAsset = useWatch({ control, name: 'sellTradeAsset' })
+  const buyTradeAsset = useWatch({ control, name: 'buyTradeAsset' })
+  const sellAssetAccount = useWatch({ control, name: 'sellAssetAccount' })
+  const amount = useWatch({ control, name: 'amount' })
+  const action = useWatch({ control, name: 'action' })
+
+  // Types
+  type TradeQuoteQueryInput = Parameters<typeof useGetTradeQuoteQuery>
+  type TradeQuoteInputArg = TradeQuoteQueryInput[0]
+
+  // State
+  const {
+    state: { wallet },
+  } = useWallet()
+  const [tradeQuoteArgs, setTradeQuoteArgs] = useState<TradeQuoteInputArg>(skipToken)
+
+  // Constants
+  const sellAsset = sellTradeAsset?.asset
+  const buyAsset = buyTradeAsset?.asset
+
+  // Selectors
+  const selectedCurrencyToUsdRate = useAppSelector(selectFiatToUsdRate)
+  const accountSpecifiersList = useSelector(selectAccountSpecifiers)
+
+  // API
+  const { data: tradeQuote } = useGetTradeQuoteQuery(tradeQuoteArgs, { pollingInterval: 30000 })
+
+  // Effects
+  // Trigger trade quote query
+  useEffect(() => {
+    const sellTradeAssetAmount = sellTradeAsset?.amount
+    if (sellAsset && buyAsset && wallet && sellTradeAssetAmount) {
+      ;(async () => {
+        const { chainId: receiveAddressChainId } = fromAssetId(buyAsset.assetId)
+        const chainAdapter = getChainAdapterManager().get(receiveAddressChainId)
+
+        if (!chainAdapter)
+          throw new Error(`couldn't get chain adapter for ${receiveAddressChainId}`)
+
+        const receiveAddress = await getFirstReceiveAddress({
+          accountSpecifiersList,
+          buyAsset,
+          chainAdapter,
+          wallet,
+        })
+        const tradeQuoteInputArgs: GetTradeQuoteInput | undefined = await (async () => {
+          const tradeQuoteInputCommonArgs: TradeQuoteInputCommonArgs = {
+            sellAmount: sellTradeAssetAmount,
+            sellAsset,
+            buyAsset,
+            sendMax: false,
+            sellAssetAccountNumber: 0,
+            receiveAddress,
+          }
+          if (isSupportedNoneUtxoSwappingChain(sellAsset?.chainId)) {
+            return {
+              ...tradeQuoteInputCommonArgs,
+              chainId: sellAsset.chainId,
+            }
+          } else if (isSupportedUtxoSwappingChain(sellAsset?.chainId) && sellAssetAccount) {
+            const { accountType, utxoParams } = getUtxoParams(sellAssetAccount)
+            if (!utxoParams?.bip44Params) throw new Error('no bip44Params')
+            const sellAssetChainAdapter = getChainAdapterManager().get(
+              sellAsset.chainId,
+            ) as unknown as UtxoBaseAdapter<UtxoSupportedChainIds>
+            const { xpub } = await sellAssetChainAdapter.getPublicKey(
+              wallet,
+              utxoParams.bip44Params,
+              accountType,
+            )
+            return {
+              ...tradeQuoteInputCommonArgs,
+              chainId: sellAsset.chainId,
+              bip44Params: utxoParams.bip44Params,
+              accountType,
+              xpub,
+            }
+          }
+        })()
+        tradeQuoteInputArgs && setTradeQuoteArgs(tradeQuoteInputArgs)
+      })()
+    }
+  }, [
+    accountSpecifiersList,
+    action,
+    amount,
+    buyAsset,
+    buyTradeAsset,
+    selectedCurrencyToUsdRate,
+    sellAsset,
+    sellAssetAccount,
+    sellTradeAsset,
+    setValue,
+    wallet,
+  ])
+
+  // Set trade quote
+  useEffect(() => {
+    tradeQuote && setValue('quote', tradeQuote)
+  }, [tradeQuote, setValue])
+}
